requirements.txt
Django>=5.0
gunicorn>=22.0
google-api-python-client>=2.127
google-auth>=2.30
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
python-dotenv>=1.0 
django-environ==django-environ-0.12.0
Pillow==11.2.1
python-decouple==3.8
psycopg2==2.9.10
supabase==2.15.2